Add files via upload
Just a little code I wrote to test your potentiometers and to find the analog values (ie 1023 is 5v). The pins on the pot are connected to the Arduino via female jumper wires. The middle is for any analog pin of your choice, and the left and right pins are for ground and 5 volts, although it doesn't matter ;)